Handle ephemeral models (#3)
## Summary and Motivation

If any of the dbt models is ephemeral, dbt2looker fails with a parsing
exception. This is because `DbtModel.relation_name` is not optional, but
it is missing for ephemeral models as they are not materialized. This PR
fixes this by making `relation_name` optional at the DbtModel class
level, but enforcing it when selecting models for looker view
generation.

## QA

I tested the new code on a manifest which contained ephemeral models.